+++ packages/docs/src/pages/datepicker/samples/DatePicker1.js
@@ -3,7 +3,7 @@ import {DatePicker, Notification} from 'react-windy-ui';
 
 export default function DatePicker1() {
   const [date, setDate] = useState();
-  const change = (text, date) => {
+  const notify = (text, date) => {
     Notification.info(
       `textFormat=${text}, date=${date}`);
     setDate(text);
@@ -11,13 +11,10 @@ export default function DatePicker1() {
 
   return <>
     <div className="doc doc-row space">
-      <DatePicker onChange={change} value={date}/>
+      <DatePicker onChange={notify} value={date} placeholder="年-月-日"/>
     </div>
     <div className="doc doc-row space">
       <DatePicker defaultValue="2018-10-11" onChange={(dateString) => console.log(dateString)}/>
-    </div>
-    <div className="doc doc-row space">
-      <DatePicker type="dateTime"/>
     </div>
   </>;
 }