+++ packages/docs/src/pages/datepicker/samples/DatePicker2.js
@@ -6,18 +6,22 @@ export default function DatePicker2() {
   const placeholder = {
     date: '请选择日期',
     month: '请选择月份',
-    year: '请选择年份'
+    year: '请选择年份',
+    dateTime: '请选择日期时间'
   }
 
-  return <InputGroup size="small" style={{maxWidth: '60%'}}>
-    <InputGroup.Item autoScale={false}>
-      <Select defaultValue={type} style={{width: '7rem'}}
-              onSelect={(value) => setType(value)}>
-        <Select.Option value="date">date</Select.Option>
-        <Select.Option value="month">month</Select.Option>
-        <Select.Option value="year">year</Select.Option>
-      </Select>
-    </InputGroup.Item>
-    <DatePicker type={type} placeholder={placeholder[type]}/>
-  </InputGroup>;
+  return <div className="doc doc-row space">
+    <InputGroup size="small" style={{maxWidth: '60%', minWidth: '20rem'}}>
+      <InputGroup.Item autoScale={false}>
+        <Select defaultValue={type} style={{width: '7rem'}}
+                onSelect={(value) => setType(value)}>
+          <Select.Option value="date">date</Select.Option>
+          <Select.Option value="month">month</Select.Option>
+          <Select.Option value="year">year</Select.Option>
+          <Select.Option value="dateTime">dateTime</Select.Option>
+        </Select>
+      </InputGroup.Item>
+      <DatePicker type={type} placeholder={placeholder[type]}/>
+    </InputGroup>
+  </div>;
 }