+++ packages/docs/src/pages/datepicker/samples/DatePicker5.js
@@ -1,9 +1,10 @@
 import React, {useState} from 'react';
-import {DatePicker, Notification, TimePicker, Toggle} from 'react-windy-ui';
+import {DatePicker, Notification, Toggle} from 'react-windy-ui';
 
 export default function DatePicker5() {
   const [hasTitle, set] = useState(false);
-  const [autoClose, setAutoClose] = useState(false);
+  const [hasFooter, enableFooter] = useState(false);
+  const [disabled, setDisabled] = useState(false);
 
   const change = (textFormat, date) => {
     Notification.info(
@@ -13,14 +14,19 @@ export default function DatePicker5() {
   return <>
     <div className="doc doc-row">
       <Toggle active={hasTitle} onChange={val => set(val)}
-              label={{on: 'With title', off: 'Without title'}}/>
+              label='With title'/>
     </div>
     <div className="doc doc-row">
-      <Toggle active={autoClose} onChange={val => setAutoClose(val)}
-              label='Auto Close'/>
+      <Toggle active={hasFooter} onChange={val => enableFooter(val)}
+              label='With footer'/>
+    </div>
+    <div className="doc doc-row">
+      <Toggle active={disabled} onChange={val => setDisabled(val)}
+              label='Disabled'/>
     </div>
     <div className="doc doc-row space">
-      <DatePicker onChange={change} type="date" hasTitle={hasTitle} autoClose={autoClose}/>
+      <DatePicker onChange={change} type="date" hasTitle={hasTitle}
+                  hasFooter={hasFooter} disabled={disabled}/>
     </div>
   </>;
 }